feat(beam/sdks/go): adding basic xlang capability and a wordcount example
