--- sdks/go/examples/cross_language/in
@@ -1,14 +1,0 @@
-1
-2
-3
-4
-4
-3
-2
-12
-1
-2
-6
-6
-7
-34
