--- sdks/go/examples/cross_language/output
@@ -1,158 +1,0 @@
-Quisque: 1
-hendrerit: 5
-fermentum: 5
-Suspendisse: 2
-sollicitudin: 2
-commodo: 2
-volutpat: 1
-Integer: 3
-semper: 3
-sapien: 1
-turpis: 7
-quis: 9
-eget: 5
-malesuada: 2
-Aenean: 2
-lectus: 2
-elementum: 3
-consectetur: 1
-at: 8
-senectus: 1
-fames: 1
-ultrices: 3
-eleifend: 6
-Proin: 2
-vel: 7
-netus: 1
-pulvinar: 1
-enim: 2
-ornare: 5
-purus: 1
-condimentum: 4
-Nullam: 2
-ipsum: 4
-accumsan: 2
-nulla: 2
-consequat: 7
-vestibulum: 3
-Fusce: 1
-vitae: 2
-nisl: 6
-congue: 5
-erat: 2
-posuere: 2
-varius: 2
-porttitor: 1
-Vestibulum: 1
-Lorem: 1
-dolor: 3
-amet: 5
-vehicula: 3
-in: 4
-odio: 3
-efficitur: 1
-tempor: 1
-tellus: 2
-Duis: 2
-rhoncus: 2
-felis: 4
-mi: 3
-Praesent: 2
-Cras: 2
-sem: 1
-feugiat: 2
-dapibus: 1
-dictum: 1
-augue: 1
-adipiscing: 1
-bibendum: 3
-suscipit: 2
-interdum: 6
-In: 3
-aliquet: 2
-luctus: 3
-eros: 7
-ullamcorper: 5
-pellentesque: 1
-non: 5
-Ut: 6
-Donec: 3
-orci: 3
-magna: 2
-metus: 7
-Mauris: 3
-rutrum: 3
-ex: 1
-auctor: 3
-lorem: 6
-laoreet: 4
-tortor: 4
-velit: 5
-ante: 2
-Nam: 1
-nisi: 2
-Nulla: 5
-habitant: 1
-mauris: 2
-Etiam: 2
-Phasellus: 1
-pharetra: 3
-iaculis: 4
-mollis: 3
-ut: 10
-risus: 4
-Curabitur: 5
-ultricies: 3
-nibh: 2
-est: 3
-eu: 6
-massa: 3
-convallis: 3
-molestie: 1
-egestas: 4
-Nunc: 2
-et: 7
-quam: 3
-faucibus: 1
-ligula: 3
-neque: 2
-venenatis: 4
-Sed: 5
-vulputate: 3
-morbi: 1
-tristique: 5
-sagittis: 1
-Maecenas: 1
-id: 11
-arcu: 4
-urna: 5
-finibus: 1
-fringilla: 2
-a: 9
-nunc: 5
-porta: 4
-ac: 7
-diam: 4
-scelerisque: 1
-leo: 4
-justo: 4
-tempus: 4
-lacus: 6
-Vivamus: 1
-mattis: 1
-nec: 1
-euismod: 3
-Aliquam: 1
-pretium: 1
-lobortis: 3
-blandit: 1
-lacinia: 1
-gravida: 2
-dui: 3
-Pellentesque: 4
-sit: 5
-tincidunt: 8
-facilisis: 1
-placerat: 3
-elit: 6
-sed: 4
