+++ sdks/go/examples/xlang/wordcount/input
@@ -1,9 +1,5 @@
 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Mauris id tellus vehicula, rutrum turpis quis, suscipit est. Quisque vehicula nec ex a interdum. Phasellus vulputate nunc sit amet nisl dapibus tincidunt ut ullamcorper nisi. Mauris gravida porta leo vel congue. Duis sit amet arcu eu nisl pharetra interdum a eget enim. Nulla facilisis massa ut egestas interdum. Nunc elit dui, hendrerit at pharetra a, pellentesque non turpis. Integer auctor vulputate congue. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Ut sagittis convallis lorem non semper. Ut ultrices elit a enim pulvinar fermentum.
-
 In rhoncus, diam sit amet laoreet ullamcorper, turpis lorem ornare tortor, eget eleifend purus risus id justo. Ut eget tortor vel elit aliquet egestas. Integer iaculis ipsum at nunc condimentum accumsan. Ut tristique felis ut metus tincidunt, quis ullamcorper diam convallis. Donec mattis ultrices lorem, id placerat tellus venenatis at. Donec quis nulla dui. Pellentesque at semper nunc. Aenean orci dui, dictum id urna a, luctus consequat augue. Nulla hendrerit mi ut quam iaculis euismod. Praesent eu velit fermentum, luctus velit id, condimentum velit.
-
 Suspendisse tempus vestibulum magna ac sollicitudin. Pellentesque id consequat lorem. Curabitur laoreet at velit a laoreet. Donec justo lectus, elementum eu elit a, placerat elementum turpis. Aliquam pretium ipsum eros, quis ultricies metus lobortis ut. Cras porta congue luctus. Curabitur vestibulum lacus est, quis eleifend metus euismod id. Duis vel ante ipsum. Proin sed posuere nulla. Sed nisi mauris, consequat ut eros vel, mollis bibendum dolor. Aenean laoreet lacus a eros iaculis eleifend.
-
 Fusce tempor tortor vel eleifend ornare. Maecenas euismod vitae nunc vel congue. Sed in tristique felis, at venenatis arcu. Etiam egestas sem quis accumsan aliquet. Ut arcu lorem, auctor et metus venenatis, ullamcorper pharetra turpis. Ut gravida, eros ac tristique faucibus, lorem quam sollicitudin lacus, tempus porta tortor risus non lectus. Nunc consequat in magna sed tincidunt. Nullam lacus mi, vulputate et erat eu, rutrum fermentum leo. Sed blandit lobortis nisl et auctor. Curabitur rutrum semper justo, quis commodo lacus tincidunt eget. Sed tempus velit ac malesuada finibus. Integer feugiat quam in metus elementum varius. Nam odio elit, tempus scelerisque est sed, tincidunt consequat ligula. Proin ligula neque, ornare et turpis at, hendrerit volutpat urna. Vivamus vel eleifend urna, quis fermentum dolor.
-
 Mauris felis urna, tincidunt quis fermentum ut, consequat eu mauris. Nulla placerat venenatis molestie. Suspendisse vitae bibendum ante. Nulla lacinia hendrerit diam non feugiat. Curabitur efficitur risus in porttitor condimentum. Pellentesque tincidunt tincidunt diam, et mollis nibh consequat id. Nulla ultrices, ligula interdum convallis varius, mi odio posuere metus, id congue risus nisl at erat. Sed rhoncus, eros eget ullamcorper interdum, leo nibh condimentum neque, eu eleifend metus odio sed justo. Etiam ultricies suscipit sapien ut ornare. Praesent ultricies fringilla nisl ac fringilla. In mollis commodo massa ac bibendum. Vestibulum interdum massa ut urna ornare iaculis. Nullam lacus eros, vehicula id hendrerit sit amet, porta a nisl. Curabitur ac leo orci. In id arcu tristique, vestibulum felis non, lobortis orci. Cras ut nunc eros.
