+++ sdks/go/examples/xlang/wordcount/output
@@ -1,0 +1,158 @@
+Lorem: 1
+ipsum: 4
+dolor: 3
+sit: 5
+amet: 5
+consectetur: 1
+adipiscing: 1
+elit: 6
+Mauris: 3
+id: 11
+tellus: 2
+vehicula: 3
+rutrum: 3
+turpis: 7
+quis: 9
+suscipit: 2
+est: 3
+Quisque: 1
+nec: 1
+ex: 1
+a: 9
+interdum: 6
+Phasellus: 1
+vulputate: 3
+nunc: 5
+nisl: 6
+dapibus: 1
+tincidunt: 8
+ut: 10
+ullamcorper: 5
+nisi: 2
+gravida: 2
+porta: 4
+leo: 4
+vel: 7
+congue: 5
+Duis: 2
+arcu: 4
+eu: 6
+pharetra: 3
+eget: 5
+enim: 2
+Nulla: 5
+facilisis: 1
+massa: 3
+egestas: 4
+Nunc: 2
+dui: 3
+hendrerit: 5
+at: 8
+pellentesque: 1
+non: 5
+Integer: 3
+auctor: 3
+Pellentesque: 4
+habitant: 1
+morbi: 1
+tristique: 5
+senectus: 1
+et: 7
+netus: 1
+malesuada: 2
+fames: 1
+ac: 7
+Ut: 6
+sagittis: 1
+convallis: 3
+lorem: 6
+semper: 3
+ultrices: 3
+pulvinar: 1
+fermentum: 5
+In: 3
+rhoncus: 2
+diam: 4
+laoreet: 4
+ornare: 5
+tortor: 4
+eleifend: 6
+purus: 1
+risus: 4
+justo: 4
+aliquet: 2
+iaculis: 4
+condimentum: 4
+accumsan: 2
+felis: 4
+metus: 7
+Donec: 3
+mattis: 1
+placerat: 3
+venenatis: 4
+nulla: 2
+Aenean: 2
+orci: 3
+dictum: 1
+urna: 5
+luctus: 3
+consequat: 7
+augue: 1
+mi: 3
+quam: 3
+euismod: 3
+Praesent: 2
+velit: 5
+Suspendisse: 2
+tempus: 4
+vestibulum: 3
+magna: 2
+sollicitudin: 2
+Curabitur: 5
+lectus: 2
+elementum: 3
+Aliquam: 1
+pretium: 1
+eros: 7
+ultricies: 3
+lobortis: 3
+Cras: 2
+lacus: 6
+ante: 2
+Proin: 2
+sed: 4
+posuere: 2
+Sed: 5
+mauris: 2
+mollis: 3
+bibendum: 3
+Fusce: 1
+tempor: 1
+Maecenas: 1
+vitae: 2
+in: 4
+Etiam: 2
+sem: 1
+faucibus: 1
+Nullam: 2
+erat: 2
+blandit: 1
+commodo: 2
+finibus: 1
+feugiat: 2
+varius: 2
+Nam: 1
+odio: 3
+scelerisque: 1
+ligula: 3
+neque: 2
+volutpat: 1
+Vivamus: 1
+molestie: 1
+lacinia: 1
+efficitur: 1
+porttitor: 1
+nibh: 2
+sapien: 1
+fringilla: 2
+Vestibulum: 1
